refactor(moderation): rename http.ts to moderationFetch.ts
